feat : Ticket 클래스에서 자동 번호 생성 로직 분리(SRP), 자동 번호 생성 로직 정적 팩토리 메서드로 변경
